refactor(widget): move error widget back to its own place
