+++ components/infobox/commons/infobox_widget.lua
@@ -7,7 +7,6 @@
 --
 local Class = require('Module:Class')
 local Lua = require('Module:Lua')
-local String = require('Module:StringUtils')
 
 local Injector = Lua.import('Module:Infobox/Widget/Injector')
 
@@ -15,11 +14,6 @@ local Injector = Lua.import('Module:Infobox/Widget/Injector')
 ---@operator call(): Widget
 ---@field public injector WidgetInjector?
 local Widget = Class.new()
-
-local _ERROR_TEXT = '<span style="color:#ff0000;font-weight:bold" class="show-when-logged-in">' ..
-					'Unexpected Error, report this in #bugs on our [https://discord.gg/liquipedia Discord]. ' ..
-					'${errorMessage}' ..
-					'</span>[[Category:Pages with script errors]]'
 
 ---Asserts the existence of a value and copies it
 ---@param value string
@@ -48,7 +42,8 @@ function Widget:tryMake()
 			mw.logObject(errorMessage, 'error')
 			mw.logObject(self, 'widget')
 			mw.log(debug.traceback())
-			return {Widget.Error({errorMessage = errorMessage})}
+			local ErrorWidget = Lua.import('Module:Infobox/Widget/Error')
+			return {ErrorWidget({errorMessage = errorMessage})}
 		end
 	)
 
@@ -63,29 +58,5 @@ function Widget:setContext(context)
 		error('Valid Injector from Infobox/Widget/Injector needs to be provided')
 	end
 end
-
--- error widget for displaying errors of widgets
--- have to add it here instead of a sep. module to avoid circular requires
-local ErrorWidget = Class.new(
-	Widget,
-	function(self, input)
-		self.errorMessage = input.errorMessage
-	end
-)
-
-function ErrorWidget:make()
-	return {
-		ErrorWidget:_create(self.errorMessage)
-	}
-end
-
-function ErrorWidget:_create(errorMessage)
-	local errorOutput = String.interpolate(_ERROR_TEXT, {errorMessage = errorMessage})
-	local errorDiv = mw.html.create('div'):node(errorOutput)
-
-	return mw.html.create('div'):node(errorDiv)
-end
-
-Widget.Error = ErrorWidget
 
 return Widget
